在gateway中，对waf、email、site、request的所有接口进行了检查及优化。 在scheduler中，对waf、image、docker的所有接口进行了检查和优化，对scheduler进行了整理和重写

+++ gateway/src/main/java/com/maoxian/gateway/mapper/RequestRecordMapper.java
@@ -43,7 +43,7 @@ public interface RequestRecordMapper {
      * @param id id
      * @return 请求模式
      */
-    Integer selectModeById(Integer id);
+    Integer selectModeById(Long id);
 
     /**
      * 查询指定时间之后的所有请求的请求时间
